[ADD] New fields en project tasks and refactor common features
