+++ somenergia_custom/views/project_view.xml
@@ -36,11 +36,35 @@
             <field name="model">project.task</field>
             <field name="inherit_id" ref="project.view_task_form2"/>
             <field name="arch" type="xml">
+                <xpath expr="//field[@name='company_id']" position="after">
+                    <field name="som_project_area_domain_ids"
+                           invisible="1"/>
+                    <field name="som_additional_project_domain_ids"
+                           invisible="1"/>
+                </xpath>
                 <xpath expr="//field[@name='tag_ids']" position="attributes">
                     <attribute name="domain">
                         [('som_for_internal_project', '=', False)]
                     </attribute>
                 </xpath>
+
+                <xpath expr="//field[@name='user_ids']" position="after">
+                    <field name="som_department_ids" widget="many2many_tags"/>
+                    <field name="som_project_id"
+                           options="{'no_open': True}"
+                           string="Area"
+                           domain="[('id', 'in', som_project_area_domain_ids)]"
+                           readonly="0"/>
+                </xpath>
+
+                <xpath expr="//field[@name='tag_ids']" position="after">
+                    <field name="som_additional_project_id"
+                           options="{'no_open': True}"
+                           string="Project"
+                           domain="[('id', 'in', som_additional_project_domain_ids)]"
+                           readonly="0"/>
+                </xpath>
+
             </field>
         </record>
